Move single file download logic to file manager
